requirements.txt
flask~=2.1.1
werkzeug~=2.0.2
wtforms~=3.0.1
pillow~=9.1.0
setuptools~=58.0.4
flask-wtf~=1.0.1
gunicorn~=20.1.0
